+++ src/lib/index.ts
@@ -1,68 +1,7 @@
-export { Type, JsonSerializer, VersionedType } from "./types";
+import { Incident } from "incident";
 
-export { ArrayType } from "./types/array";
-export { BooleanType } from "./types/boolean";
-export { BufferType } from "./types/buffer";
-export { CodepointStringType } from "./types/codepoint-string";
-export { CustomType } from "./types/custom";
-export { DateType } from "./types/date";
-export { DocumentType } from "./types/document";
-export { Float64Type } from "./types/float64";
-export { IntegerType } from "./types/integer";
-export { JsonType } from "./types/json";
-export { LiteralType } from "./types/literal";
-export { MapType } from "./types/map";
-export { NullType } from "./types/null";
-export { SimpleEnumType } from "./types/simple-enum";
-export { TaggedUnionType } from "./types/tagged-union";
-export { TryUnionType } from "./types/try-union";
-export { WhiteListType } from "./types/white-list";
-export { Ucs2StringType } from "./types/ucs2-string";
-export { UnionType } from "./types/union";
-
-import * as errors from "./errors/index";
-import * as array from "./types/array";
-import * as boolean from "./types/boolean";
-import * as buffer from "./types/buffer";
-import * as codepointString from "./types/codepoint-string";
-import * as custom from "./types/custom";
-import * as date from "./types/date";
-import * as document from "./types/document";
-import * as float64 from "./types/float64";
-import * as integer from "./types/integer";
-import * as json from "./types/json";
-import * as literal from "./types/literal";
-import * as map from "./types/map";
-import * as _null from "./types/null";
-import * as simpleEnum from "./types/simple-enum";
-import * as taggedUnion from "./types/tagged-union";
-import * as tryUnion from "./types/try-union";
-import * as ucs2String from "./types/ucs2-string";
-import * as union from "./types/union";
-import * as whiteList from "./types/white-list";
-
-export {
-  array,
-  boolean,
-  buffer,
-  codepointString,
-  custom,
-  date,
-  document,
-  float64,
-  integer,
-  literal,
-  json,
-  map,
-  _null as null,
-  simpleEnum,
-  taggedUnion,
-  tryUnion,
-  whiteList,
-  ucs2String,
-  union,
-};
-
-export { errors };
-
-export { CaseStyle } from "./case-style";
+throw new Incident(
+  "IndexImport",
+  // tslint:disable-next-line:max-line-length
+  "You cannot import `kryo`'s index module, use deep imports such as `import { DocumentType } from \"kryo/types/document\"`.",
+);